improvement: change a double types to float

+++ src/main/java/com/kata/supermarket/model/IProductOrder.java
@@ -1,5 +1,5 @@
 package com.kata.supermarket.model;
 
 public interface IProductOrder {
-    double calculateProductOrderPrice();
+    float calculateProductOrderPrice();
 }
